requirements.txt
acquire-imaging==0.4.0rc2
python-dotenv
s3fs
zarr